+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.339](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.338...@codecademy/gamut-kit@0.6.339) (2023-08-04)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.6.338](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.337...@codecademy/gamut-kit@0.6.338) (2023-08-02)
 
 **Note:** Version bump only for package @codecademy/gamut-kit